refactor(ai-diagnostic): split diagnostic_service monolith into cohesive modules
diagnostic_service.py was a ~2800-line monolith with no direct test coverage.
Added 20 characterisation tests over its pure seams (data models, rule-based
backend, learning engine, cloud-client mapping helpers), then extracted those
seams into focused modules under the green suite:

  service/models.py          - Problem, Solution, LearnedPattern
  service/learning_engine.py - LearningEngine (feedback-driven confidence)
  service/backends.py        - AIBackend, RuleBasedBackend, OllamaBackend + RULES
  service/cloud_client.py    - CloudClient (solution -> cloud command)

diagnostic_service.py re-exports each name, so existing importers and the
runtime entrypoint are unchanged. 2807 -> 2150 LOC in the monolith. All 173
ai-diagnostic tests pass (153 existing + 20 new); the learning-confidence
magic numbers are now named constants (behaviour identical).
