+++ ai-diagnostic/service/backends.py
@@ -1,0 +1,398 @@
+"""
+AI diagnostic backends.
+
+An `AIBackend` turns a `Problem` into a `Solution`. Two implementations are
+provided: a fully offline rule-based expert system (`RuleBasedBackend`, the
+default, no network needed) and a local-LLM backend (`OllamaBackend`) that falls
+back to the rule-based one when the model is unreachable. Extracted from
+diagnostic_service.py; the large static RULES catalogue lives with the backend
+that consumes it.
+"""
+
+import json
+import logging
+from abc import ABC, abstractmethod
+
+from service.models import Problem, Solution
+
+logger = logging.getLogger(__name__)
+
+
+class AIBackend(ABC):
+    """Abstract base class for AI diagnostic backends"""
+
+    @abstractmethod
+    def diagnose(self, problem: Problem) -> Solution:
+        """Analyze problem and return solution"""
+        pass
+
+    @abstractmethod
+    def is_available(self) -> bool:
+        """Check if backend is available"""
+        pass
+
+
+class RuleBasedBackend(AIBackend):
+    """
+    Rule-based expert system for offline operation.
+    Works without any API keys or internet connection.
+    """
+
+    RULES = {
+        "CPU_OVERHEAT": {
+            "action": "Reduce thermal load and increase cooling",
+            "commands": [
+                "echo 1 > /sys/class/thermal/cooling_device0/cur_state",
+                "systemctl restart fan-controller",
+                "cpufreq-set -g powersave",
+                "kill -STOP $(pgrep -f non-critical)"
+            ],
+            "expected_outcome": "CPU temperature should drop below 70C within 5 minutes",
+            "risk_level": "low"
+        },
+        "MEMORY_PRESSURE": {
+            "action": "Free memory and restart memory-heavy processes",
+            "commands": [
+                "sync; echo 3 > /proc/sys/vm/drop_caches",
+                "systemctl restart radio_daemon",
+                "pkill -9 -f memory_leak_process",
+                "swapon -a"
+            ],
+            "expected_outcome": "Memory usage should drop below 70%",
+            "risk_level": "low"
+        },
+        "SIGNAL_DEGRADATION": {
+            "action": "Optimize antenna and adjust transmission parameters",
+            "commands": [
+                "radio-cli recalibrate-antenna",
+                "radio-cli set-power auto",
+                "radio-cli scan-interference",
+                "systemctl restart radio_controller"
+            ],
+            "expected_outcome": "Signal strength should improve by 10-15 dBm",
+            "risk_level": "low"
+        },
+        "BACKHAUL_LATENCY": {
+            "action": "Optimize network path and reduce traffic",
+            "commands": [
+                "tc qdisc replace dev eth0 root fq_codel",
+                "ip route flush cache",
+                "systemctl restart network-optimizer",
+                "ethtool -s eth0 speed 1000 duplex full"
+            ],
+            "expected_outcome": "Latency should drop below 50ms",
+            "risk_level": "low"
+        },
+        "PROCESS_CRASH": {
+            "action": "Restart crashed process and enable core dumps for analysis",
+            "commands": [
+                "ulimit -c unlimited",
+                "systemctl restart radio_controller",
+                "journalctl -u radio_controller --since '5 min ago' > /var/log/crash_analysis.log",
+                "coredumpctl gdb radio_controller"
+            ],
+            "expected_outcome": "Process should restart and remain stable",
+            "risk_level": "medium"
+        },
+        "CONFIG_MISMATCH": {
+            "action": "Restore configuration from backup and validate",
+            "commands": [
+                "cp /etc/basestation/radio.conf.backup /etc/basestation/radio.conf",
+                "config-validator --check /etc/basestation/radio.conf",
+                "systemctl reload radio_controller",
+                "logger 'Config restored from backup'"
+            ],
+            "expected_outcome": "Configuration should be valid and applied",
+            "risk_level": "low"
+        },
+        "POWER_FLUCTUATION": {
+            "action": "Activate power protection and reduce load",
+            "commands": [
+                "power-manager --enable-protection",
+                "cpufreq-set -g powersave",
+                "radio-cli reduce-power 20",
+                "ups-cli --check-battery"
+            ],
+            "expected_outcome": "System should switch to UPS if needed, power stabilized",
+            "risk_level": "high"
+        },
+        "HIGH_POWER_CONSUMPTION": {
+            "action": "Reduce power consumption through efficiency optimization",
+            "commands": [
+                "radio-cli set-power-mode eco",
+                "cpufreq-set -u 1.5GHz",
+                "systemctl stop non-essential.target",
+                "power-manager --optimize"
+            ],
+            "expected_outcome": "Power consumption should drop below rated capacity",
+            "risk_level": "low"
+        },
+        "AUTH_FAILURE": {
+            "action": "Lock suspicious source and strengthen security",
+            "commands": [
+                "iptables -A INPUT -s 192.168.1.100 -j DROP",
+                "fail2ban-client set sshd banip 192.168.1.100",
+                "passwd -l admin",
+                "logger -p auth.warning 'Brute force attack detected'"
+            ],
+            "expected_outcome": "Attack source blocked, accounts secured",
+            "risk_level": "low"
+        },
+        "CERT_EXPIRING": {
+            "action": "Renew TLS certificate",
+            "commands": [
+                "certbot renew --force-renewal",
+                "systemctl reload nginx",
+                "openssl x509 -in /etc/ssl/certs/server.crt -noout -dates",
+                "logger 'Certificate renewed'"
+            ],
+            "expected_outcome": "Certificate renewed with new expiry date",
+            "risk_level": "low"
+        },
+        # 5G NR Specific Rules (based on Huawei SSV criteria)
+        "TX_IMBALANCE_HIGH": {
+            "action": "Investigate RF path imbalance and recalibrate antenna system",
+            "commands": [
+                "radio-cli check-tx-path --all-sectors",
+                "radio-cli measure-vswr --threshold 1.5",
+                "radio-cli calibrate-rf-chain",
+                "radio-cli verify-antenna-connections",
+                "logger -p local0.alert 'TX imbalance exceeded 4dB threshold - SSV FAIL'"
+            ],
+            "expected_outcome": "TX imbalance should drop below 4dB threshold (SSV pass criteria)",
+            "risk_level": "high"
+        },
+        "DL_THROUGHPUT_LOW_NR3500": {
+            "action": "Optimize NR3500 downlink performance",
+            "commands": [
+                "radio-cli check-interference --band NR3500",
+                "radio-cli optimize-mcs --target 26",
+                "radio-cli adjust-power --band NR3500 --mode auto",
+                "radio-cli verify-backhaul-capacity",
+                "logger 'NR3500 DL throughput below 1000 Mbps KPI threshold'"
+            ],
+            "expected_outcome": "DL throughput should reach >= 1000 Mbps (NR3500 100MHz RANK4 256QAM)",
+            "risk_level": "medium"
+        },
+        "UL_THROUGHPUT_LOW_NR3500": {
+            "action": "Optimize NR3500 uplink performance",
+            "commands": [
+                "radio-cli check-ul-interference --band NR3500",
+                "radio-cli optimize-ul-grant",
+                "radio-cli adjust-ul-power --band NR3500",
+                "radio-cli verify-ul-sync",
+                "logger 'NR3500 UL throughput below 75 Mbps KPI threshold'"
+            ],
+            "expected_outcome": "UL throughput should reach >= 75 Mbps (NR3500 100MHz RANK1 256QAM)",
+            "risk_level": "medium"
+        },
+        "DL_THROUGHPUT_LOW_NR700": {
+            "action": "Optimize NR700 coverage layer performance",
+            "commands": [
+                "radio-cli check-interference --band NR700",
+                "radio-cli optimize-coverage-layer",
+                "radio-cli adjust-power --band NR700 --mode coverage",
+                "logger 'NR700 DL throughput below 50 Mbps KPI threshold'"
+            ],
+            "expected_outcome": "DL throughput should reach >= 50 Mbps (NR700 10MHz RANK2 256QAM)",
+            "risk_level": "low"
+        },
+        "LATENCY_HIGH": {
+            "action": "Reduce 5G air interface and backhaul latency",
+            "commands": [
+                "radio-cli optimize-scheduling --low-latency",
+                "radio-cli check-harq-timing",
+                "tc qdisc replace dev eth0 root fq_codel target 5ms",
+                "radio-cli verify-fronthaul-latency",
+                "logger 'Latency exceeded 15ms 5G target'"
+            ],
+            "expected_outcome": "Latency should drop below 15ms (5G target)",
+            "risk_level": "medium"
+        },
+        "SINR_DEGRADATION": {
+            "action": "Improve signal-to-interference ratio",
+            "commands": [
+                "radio-cli scan-interference --detailed",
+                "radio-cli optimize-beamforming",
+                "radio-cli adjust-tilt --optimize-sinr",
+                "radio-cli check-pci-collision",
+                "logger 'SINR below 10dB - coverage degradation detected'"
+            ],
+            "expected_outcome": "SINR should improve above 15dB for good coverage",
+            "risk_level": "medium"
+        },
+        "HANDOVER_FAILURE": {
+            "action": "Investigate and fix inter-cell handover issues",
+            "commands": [
+                "radio-cli check-neighbor-relations",
+                "radio-cli verify-x2-connectivity",
+                "radio-cli analyze-handover-logs --last 100",
+                "radio-cli optimize-handover-params",
+                "logger -p local0.alert 'Handover success rate below 100% - SSV FAIL'"
+            ],
+            "expected_outcome": "Handover success rate should reach 100% (SSV criteria)",
+            "risk_level": "low"
+        },
+        "RSRP_WEAK": {
+            "action": "Improve reference signal coverage",
+            "commands": [
+                "radio-cli increase-rs-power",
+                "radio-cli optimize-antenna-tilt",
+                "radio-cli check-feeder-loss",
+                "radio-cli verify-antenna-gain",
+                "logger 'RSRP below -100dBm - weak coverage area'"
+            ],
+            "expected_outcome": "RSRP should improve above -85dBm for good coverage",
+            "risk_level": "low"
+        },
+        "BLER_HIGH": {
+            "action": "Reduce block error rate",
+            "commands": [
+                "radio-cli analyze-bler-distribution",
+                "radio-cli optimize-harq-retx",
+                "radio-cli adjust-mcs-table",
+                "radio-cli check-timing-advance",
+                "logger 'Initial BLER exceeding 10% threshold'"
+            ],
+            "expected_outcome": "BLER should drop below 10% for stable transmission",
+            "risk_level": "medium"
+        },
+        # Extended problem codes for comprehensive AI diagnostics
+        "HIGH_BLOCK_ERROR_RATE": {
+            "action": "Reduce block error rate through RF optimization",
+            "commands": [
+                "radio-cli analyze-bler-distribution",
+                "radio-cli optimize-harq-retx --max-retx 4",
+                "radio-cli adjust-mcs-table --conservative",
+                "radio-cli check-timing-advance",
+                "radio-cli scan-interference --all-bands"
+            ],
+            "expected_outcome": "BLER should drop below 15% warning threshold",
+            "risk_level": "low"
+        },
+        "LOW_BATTERY": {
+            "action": "Activate power saving mode and prepare for failover",
+            "commands": [
+                "power-manager --mode emergency-save",
+                "radio-cli reduce-power 50",
+                "systemctl stop non-essential.target",
+                "ups-cli --check-generator",
+                "logger -p local0.crit 'Battery critically low - emergency mode activated'"
+            ],
+            "expected_outcome": "Power consumption reduced, generator standby activated",
+            "risk_level": "medium"
+        },
+        "HIGH_LATENCY": {
+            "action": "Optimize network path and reduce latency",
+            "commands": [
+                "tc qdisc replace dev eth0 root fq_codel",
+                "ip route flush cache",
+                "radio-cli optimize-scheduling --low-latency",
+                "systemctl restart network-optimizer",
+                "ping -c 10 8.8.8.8 | tail -1"
+            ],
+            "expected_outcome": "Latency should drop below 50ms warning threshold",
+            "risk_level": "low"
+        },
+        "LOW_THROUGHPUT": {
+            "action": "Optimize throughput and check backhaul capacity",
+            "commands": [
+                "radio-cli check-backhaul-capacity",
+                "radio-cli optimize-mcs --target max",
+                "radio-cli check-interference --all-bands",
+                "iperf3 -c backhaul-gateway -t 10",
+                "systemctl restart traffic-shaper"
+            ],
+            "expected_outcome": "Throughput should increase above 50 Mbps threshold",
+            "risk_level": "low"
+        },
+        # Note: HANDOVER_FAILURE rule defined above at SSV-specific section (line ~705)
+        "HIGH_INTERFERENCE": {
+            "action": "Mitigate RF interference through frequency and power optimization",
+            "commands": [
+                "radio-cli scan-interference --detailed",
+                "radio-cli auto-frequency-select",
+                "radio-cli adjust-power --mode interference-aware",
+                "radio-cli enable-icic",
+                "logger 'High interference detected - mitigation active'"
+            ],
+            "expected_outcome": "Interference level should drop below -80 dBm threshold",
+            "risk_level": "low"
+        }
+    }
+
+    def diagnose(self, problem: Problem) -> Solution:
+        rule = self.RULES.get(problem.code, {
+            "action": "Manual investigation required - unknown problem type",
+            "commands": [
+                f"logger 'Unknown problem: {problem.code}'",
+                "dmesg | tail -100 > /var/log/diagnostic.log",
+                "systemctl status --all > /var/log/services.log"
+            ],
+            "expected_outcome": "Logs collected for manual analysis",
+            "risk_level": "unknown"
+        })
+
+        return Solution(
+            problem_id=problem.id,
+            action=rule["action"],
+            commands=rule["commands"],
+            expected_outcome=rule["expected_outcome"],
+            risk_level=rule["risk_level"],
+            confidence=0.92 if problem.code in self.RULES else 0.3,  # 92% for rules (above 90% auto-apply threshold)
+            reasoning=f"Rule-based diagnosis for {problem.code}"
+        )
+
+    def is_available(self) -> bool:
+        return True
+
+
+class OllamaBackend(AIBackend):
+    """Local Ollama backend for offline AI diagnosis"""
+
+    def __init__(self, model: str = "llama3.2", host: str = "http://localhost:11434"):
+        self.model = model
+        self.host = host
+
+    def diagnose(self, problem: Problem) -> Solution:
+        import requests
+
+        prompt = f"""Diagnose this base station problem and provide a solution:
+
+Problem: {problem.code} - {problem.message}
+Metrics: {json.dumps(problem.metrics)}
+Logs: {problem.raw_logs}
+
+Respond with JSON: {{"action": "...", "commands": [...], "expected_outcome": "...", "risk_level": "low|medium|high"}}"""
+
+        try:
+            response = requests.post(
+                f"{self.host}/api/generate",
+                json={"model": self.model, "prompt": prompt, "stream": False},
+                timeout=60
+            )
+
+            if response.ok:
+                text = response.json().get("response", "")
+                data = json.loads(text)
+                return Solution(
+                    problem_id=problem.id,
+                    action=data.get("action", ""),
+                    commands=data.get("commands", []),
+                    expected_outcome=data.get("expected_outcome", ""),
+                    risk_level=data.get("risk_level", "medium"),
+                    confidence=0.7,
+                    reasoning="Local LLM analysis"
+                )
+        except Exception as e:
+            logger.error(f"Ollama error: {e}")
+
+        return RuleBasedBackend().diagnose(problem)
+
+    def is_available(self) -> bool:
+        try:
+            import requests
+            r = requests.get(f"{self.host}/api/tags", timeout=2)
+            return r.ok
+        except Exception:
+            return False
