+++ ai-diagnostic/service/models.py
@@ -1,0 +1,54 @@
+"""
+Core data models for the AI diagnostic service.
+
+These are the universal, protocol-agnostic value types shared by the diagnostic
+backends, the learning engine and the cloud client. Extracted from
+diagnostic_service.py so they can be imported without dragging in the socket /
+HTTP / MQTT machinery.
+"""
+
+from dataclasses import dataclass, field
+from typing import Any, Dict, List
+
+
+@dataclass
+class Problem:
+    """Universal problem format from any device"""
+    id: str
+    timestamp: str
+    station_id: str
+    category: str
+    severity: str
+    code: str
+    message: str
+    metrics: Dict[str, Any]
+    raw_logs: str = ""  # Optional - may be empty from HTTP requests
+    source_protocol: str = "unknown"
+
+
+@dataclass
+class Solution:
+    """Solution generated by AI"""
+    problem_id: str
+    action: str
+    commands: List[str]
+    expected_outcome: str
+    risk_level: str
+    confidence: float = 0.0
+    reasoning: str = ""
+
+
+@dataclass
+class LearnedPattern:
+    """Pattern learned from operator feedback"""
+    problem_code: str
+    category: str
+    resolved_count: int = 0
+    failed_count: int = 0
+    adjusted_confidence: float = 0.85
+    successful_actions: List[str] = field(default_factory=list)
+    failed_actions: List[str] = field(default_factory=list)
+
+    def success_rate(self) -> float:
+        total = self.resolved_count + self.failed_count
+        return (self.resolved_count / total * 100) if total > 0 else 0.0
